+++ table/table_regress_test.go
@@ -5,14 +5,17 @@ package table_test
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
+	"golang.org/x/xerrors"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table"
+	"github.com/ydb-platform/ydb-go-sdk/v3/table/result/named"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
 )
 
@@ -114,4 +117,51 @@ func TestIssue259IntervalFromDuration(t *testing.T) {
 		return nil
 	})
 	require.NoError(t, err)
+}
+
+func TestIssue415ScanError(t *testing.T) {
+	// https://github.com/ydb-platform/ydb-go-sdk/issues/415
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	db := connect(t)
+	defer db.Close(ctx)
+	err := db.Table().DoTx(ctx, func(ctx context.Context, tx table.TransactionActor) error {
+		res, err := tx.Execute(ctx, `SELECT 1 as abc, 2 as def;`, nil)
+		if err != nil {
+			return err
+		}
+		err = res.NextResultSetErr(ctx)
+		if err != nil {
+			return err
+		}
+		if !res.NextRow() {
+			if err = res.Err(); err != nil {
+				return err
+			}
+			return fmt.Errorf("unexpected empty result set")
+		}
+		var abc, def int32
+		err = res.ScanNamed(
+			named.Required("abc", &abc),
+			named.Required("ghi", &def),
+		)
+		if err != nil {
+			return err
+		}
+		fmt.Println(abc, def)
+		return res.Err()
+	}, table.WithTxSettings(table.TxSettings(table.WithSnapshotReadOnly())))
+	require.Error(t, err)
+	err = func(err error) error {
+		for {
+			//nolint:errorlint
+			if unwrappedErr, has := err.(xerrors.Wrapper); has {
+				err = unwrappedErr.Unwrap()
+			} else {
+				return err
+			}
+		}
+	}(err)
+	require.Equal(t, "not found column 'ghi'", err.Error())
 }