+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed bug from scan unexpected column name
+
 ## v3.38.4
 * Changed type of `table/options.{Create,Alter,Drop}TableOption` from func to interface
 * Added implementations of `table/options.{Create,Alter,Drop}Option`